Design Update
Changed the logo to the agreed upon logo version.

Instead of using simple text as the title of our website I switched it to an image to have the same look as the logo.

I increased the font for the different pages to be more visible.

I made the background a light gray so its not just basic white, giving it a little bit of a crisp feel to it.